+++ History.md
@@ -1,9 +1,16 @@
-## v0.1.0
+## vNEXT (v2.0.0)
 
 * Split off the translation UI into a separate package (#29)
+* Server now keeps track of connection locales (#83) thanks @lucazulian
+
+* localStorage is now used to cache: strings and lastUpdatedAt times for
+  each language, and the user's current locale.
+
+* msgfmt.locale(), msgfmt.lang(), msgfmt.dir() reactive getters
+
+* msgfmt.setBodyDir = true (default) will set <body> direction
 
 TODO
 
 * Retrieve languages via separate JSON request, cache it
-* localStorage for lastUpdateTime of different languages?
 * Language updates via DDP.